+++ src/scripts/lib/serviceworkers.js
@@ -5,6 +5,7 @@ import rollup from "rollup";
 import uglifyjs from "uglify-js";
 import nodeResolve from "rollup-plugin-node-resolve";
 import commonjs from "rollup-plugin-commonjs";
+import rollupJSON from "rollup-plugin-json";
 import babel from "rollup-plugin-babel";
 
 export function serviceworkers(opts = {}) {
@@ -13,6 +14,7 @@ export function serviceworkers(opts = {}) {
     plugins: [
       nodeResolve(),
       commonjs(),
+      rollupJSON(),
       babel({
         exclude: "node_modules/**",
         babelrc: false,
@@ -20,25 +22,13 @@ export function serviceworkers(opts = {}) {
       })
     ]
   })
-  .then(bundle => {
-      console.log("after rollup");
-      return bundle;
-  })
   .then(bundle => bundle.generate({
     format: "cjs",
-    sourceMap: true
+    sourceMap: false
   }))
-  .then(obj => {
-      console.log("after generate");
-      return obj;
-  })
   .then(obj => opts.minify ? uglifyjs.minify(obj.code, {
     fromString: true,
     mangle: true
   }) : obj)
-  .then(obj => obj.code)
-  .then(code => {
-      console.log(code);
-      return code;
-  });
+  .then(obj => obj.code);
 }